[chore] re-enable testPrometheusAnnotationMetrics test
