+++ functional_tests/functional/functional_test.go
@@ -5,7 +5,6 @@ package functional
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -25,14 +24,11 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
-	appextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	k8sruntime "k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/runtime/serializer"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
@@ -112,147 +108,6 @@ func setupSinks(t *testing.T) {
 		k8sclusterReceiverMetricsConsumer: internal.SetupSignalfxReceiver(t,
 			signalFxReceiverK8sClusterReceiverPort),
 		tracesConsumer: internal.SetupOTLPTracesSink(t),
-	}
-}
-
-func requiresPrometheusCRD(kubeTestEnv string) bool {
-	return kubeTestEnv == kindTestKubeEnv
-}
-
-func deployPrometheusResources(t *testing.T, extensionsClient *clientset.Clientset, dynamicClient dynamic.Interface) {
-	decode := scheme.Codecs.UniversalDeserializer().Decode
-
-	// load up Prometheus PodMonitor and ServiceMonitor CRDs:
-	stream, err := os.ReadFile(filepath.Join(testDir, manifestsDir, "prometheus_operator_crds.yaml"))
-	require.NoError(t, err)
-	sch := k8sruntime.NewScheme()
-
-	var obj k8sruntime.Object
-	var groupVersionKind *schema.GroupVersionKind
-	for _, resourceYAML := range strings.Split(string(stream), "---") {
-		if len(resourceYAML) == 0 {
-			continue
-		}
-
-		obj, groupVersionKind, err = decode(
-			[]byte(resourceYAML),
-			nil,
-			nil)
-		require.NoError(t, err)
-		if groupVersionKind.Group == "apiextensions.k8s.io" &&
-			groupVersionKind.Version == "v1" &&
-			groupVersionKind.Kind == "CustomResourceDefinition" {
-			crd := obj.(*appextensionsv1.CustomResourceDefinition)
-			apiExtensions := extensionsClient.ApiextensionsV1().CustomResourceDefinitions()
-
-			existing, getErr := apiExtensions.Get(t.Context(), crd.Name, metav1.GetOptions{})
-			if getErr == nil && existing.DeletionTimestamp != nil {
-				t.Logf("CRD %s is terminating, waiting for removal...", crd.Name)
-				require.EventuallyWithT(t, func(tt *assert.CollectT) {
-					_, e := apiExtensions.Get(t.Context(), crd.Name, metav1.GetOptions{})
-					assert.True(tt, k8serrors.IsNotFound(e), "CRD %s still exists", crd.Name)
-				}, 3*time.Minute, 3*time.Second, "CRD %s stuck in Terminating", crd.Name)
-			}
-
-			crdName := crd.Name
-			crdSpec := crd.Spec
-			created, createErr := apiExtensions.Create(t.Context(), crd, metav1.CreateOptions{})
-			if createErr != nil {
-				if k8serrors.IsAlreadyExists(createErr) {
-					t.Logf("CRD %s already exists, skipping creation", crdName)
-				} else {
-					require.NoError(t, createErr)
-				}
-			} else {
-				t.Logf("Deployed CRD %s", created.Name)
-			}
-
-			require.EventuallyWithT(t, func(tt *assert.CollectT) {
-				latest, latestErr := apiExtensions.Get(t.Context(), crdName, metav1.GetOptions{})
-				assert.NoError(tt, latestErr)
-				established := false
-				for _, cond := range latest.Status.Conditions {
-					if cond.Type == appextensionsv1.Established && cond.Status == appextensionsv1.ConditionTrue {
-						established = true
-					}
-				}
-				assert.True(tt, established)
-			}, 3*time.Minute, 3*time.Second, "CRD %s not established", crdName)
-
-			for _, version := range crdSpec.Versions {
-				sch.AddKnownTypeWithName(
-					schema.GroupVersionKind{
-						Group:   crdSpec.Group,
-						Version: version.Name,
-						Kind:    crdSpec.Names.Kind,
-					},
-					&unstructured.Unstructured{},
-				)
-			}
-		}
-	}
-
-	codecs := serializer.NewCodecFactory(sch)
-	crdDecode := codecs.UniversalDeserializer().Decode
-
-	// Prometheus pod monitor
-	stream, err = os.ReadFile(filepath.Join(testDir, manifestsDir, "pod_monitor.yaml"))
-	require.NoError(t, err)
-	podMonitor, _, err := crdDecode(stream, nil, nil)
-	require.NoError(t, err)
-	g := schema.GroupVersionResource{
-		Group:    "monitoring.coreos.com",
-		Version:  "v1",
-		Resource: "podmonitors",
-	}
-	_, err = dynamicClient.Resource(g).Namespace(internal.DefaultNamespace).Create(t.Context(),
-		podMonitor.(*unstructured.Unstructured), metav1.CreateOptions{})
-	assert.NoError(t, err)
-
-	// Prometheus service monitor
-	stream, err = os.ReadFile(filepath.Join(testDir, manifestsDir, "service_monitor.yaml"))
-	require.NoError(t, err)
-	serviceMonitor, _, err := crdDecode(stream, nil, nil)
-	require.NoError(t, err)
-	g = schema.GroupVersionResource{
-		Group:    "monitoring.coreos.com",
-		Version:  "v1",
-		Resource: "servicemonitors",
-	}
-	_, err = dynamicClient.Resource(g).Namespace(internal.DefaultNamespace).Create(t.Context(),
-		serviceMonitor.(*unstructured.Unstructured), metav1.CreateOptions{})
-	assert.NoError(t, err)
-}
-
-func teardownPrometheusResources(ctx context.Context, t *testing.T, extensionsClient *clientset.Clientset) {
-	decode := scheme.Codecs.UniversalDeserializer().Decode
-	waitTime := int64(0)
-	crdstream, err := os.ReadFile(filepath.Join(testDir, manifestsDir, "prometheus_operator_crds.yaml"))
-	require.NoError(t, err)
-	var obj k8sruntime.Object
-	var groupVersionKind *schema.GroupVersionKind
-	for _, resourceYAML := range strings.Split(string(crdstream), "---") {
-		if len(resourceYAML) == 0 {
-			continue
-		}
-
-		obj, groupVersionKind, err = decode(
-			[]byte(resourceYAML),
-			nil,
-			nil)
-		require.NoError(t, err)
-		if groupVersionKind.Group == "apiextensions.k8s.io" &&
-			groupVersionKind.Version == "v1" &&
-			groupVersionKind.Kind == "CustomResourceDefinition" {
-			crd := obj.(*appextensionsv1.CustomResourceDefinition)
-			apiExtensions := extensionsClient.ApiextensionsV1().CustomResourceDefinitions()
-			err = apiExtensions.Delete(ctx, crd.Name, metav1.DeleteOptions{
-				GracePeriodSeconds: &waitTime,
-			})
-			if err != nil && !k8serrors.IsNotFound(err) {
-				t.Logf("Failed to delete CRD %s during teardown: %v", crd.Name, err)
-			}
-		}
 	}
 }
 
@@ -450,6 +305,8 @@ func teardown(ctx context.Context, t *testing.T, testKubeConfig string) {
 	require.NoError(t, err)
 	extensionsClient, err := clientset.NewForConfig(kubeConfig)
 	require.NoError(t, err)
+	dynamicClient, err := dynamic.NewForConfig(kubeConfig)
+	require.NoError(t, err)
 	waitTime := int64(0)
 	deployments := client.AppsV1().Deployments(internal.DefaultNamespace)
 	require.NoError(t, err)
@@ -515,7 +372,7 @@ func teardown(ctx context.Context, t *testing.T, testKubeConfig string) {
 	}
 
 	if requiresPrometheusCRD(os.Getenv("KUBE_TEST_ENV")) {
-		teardownPrometheusResources(ctx, t, extensionsClient)
+		teardownPrometheusResources(ctx, t, extensionsClient, dynamicClient)
 	}
 
 	for _, nm := range namespaces {
@@ -588,8 +445,7 @@ func runLocalClusterTests(t *testing.T) {
 	t.Run("test k8s objects", testK8sObjects)
 	t.Run("test agent metrics", testAgentMetrics)
 	t.Run("test target allocator", testTargetAllocator)
-	// TODO: re-enable this test in 0.129.0 https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/40788
-	// t.Run("test prometheus metrics", testPrometheusAnnotationMetrics)
+	t.Run("test prometheus metrics", testPrometheusAnnotationMetrics)
 
 	// Test component health - verify no RBAC or connection errors
 	t.Run("component error logs checks", func(t *testing.T) {
@@ -744,12 +600,6 @@ func findResourceByAttr(t *testing.T, filePath string, attrKey string, skipKeys 
 	}
 	require.Failf(t, "resource not found", "no ResourceMetrics with attribute %q in %s", attrKey, filePath)
 	return pcommon.NewMap()
-}
-
-func requireEnv(t *testing.T, key string) string {
-	value, set := os.LookupEnv(key)
-	require.True(t, set, "the environment variable %s must be set", key)
-	return value
 }
 
 func containerImageShorten(value string) string {
@@ -1000,72 +850,6 @@ func testK8sObjects(t *testing.T) {
 	assert.True(t, foundCustomField2)
 }
 
-func testTargetAllocator(t *testing.T) {
-	if !requiresPrometheusCRD(os.Getenv("KUBE_TEST_ENV")) {
-		t.Fatalf("Required Prometheus CRDs are not installed")
-	}
-
-	testKubeConfig := requireEnv(t, "KUBECONFIG")
-	kubeConfig, err := clientcmd.BuildConfigFromFlags("", testKubeConfig)
-	require.NoError(t, err)
-	client, err := kubernetes.NewForConfig(kubeConfig)
-	require.NoError(t, err)
-
-	// check target allocator logs
-	require.EventuallyWithT(t, func(c *assert.CollectT) {
-		var taPodList *corev1.PodList
-		taPodList, err = internal.GetPods(t, client, internal.DefaultNamespace, internal.TargetAllocatorLabelSelector)
-		assert.NoError(c, err)
-		containsReadyTAPod := false
-
-		for _, pod := range taPodList.Items {
-			if pod.Status.Phase != "Running" {
-				t.Logf("Skipping pod %s in phase %s", pod.Name, pod.Status.Phase)
-				continue
-			}
-			containsReadyTAPod = true
-			var podLogs string
-			podLogs, err = internal.GetPodLogs(t, client, internal.DefaultNamespace, pod.Name, internal.TargetAllocatorContainerName, 100)
-			assert.NoError(c, err)
-			assert.Contains(c, podLogs, "Service Discovery watch event received", "Target allocator pod logs failed to successfully discover targets. Received logs: %v", podLogs)
-		}
-		assert.True(c, containsReadyTAPod, "No target allocator pod found ready")
-	}, 3*time.Minute, 3*time.Second, "Failed to find required target allocator pod logs")
-
-	// check agent logs
-	serviceMonitorRegex := regexp.MustCompile(`Scrape job added.*"otelcol\.component\.id": "prometheus/ta.*"jobName": "serviceMonitor/default/prometheus-service-monitor/0"`)
-	podMonitorRegex := regexp.MustCompile(`Scrape job added.*"otelcol\.component\.id": "prometheus/ta.*"jobName": "podMonitor/default/pod-monitor/0"`)
-	require.EventuallyWithT(t, func(c *assert.CollectT) {
-		var agentPodList *corev1.PodList
-		agentPodList, err = internal.GetPods(t, client, internal.DefaultNamespace, internal.AgentLabelSelector)
-		assert.NoError(c, err)
-		containsReadyAgentPod := false
-		var combinedPodLogs strings.Builder
-
-		for i, pod := range agentPodList.Items {
-			if pod.Status.Phase != "Running" {
-				t.Logf("Skipping pod %s in phase %s", pod.Name, pod.Status.Phase)
-				continue
-			}
-			containsReadyAgentPod = true
-			var podLogs string
-			podLogs, err = internal.GetPodLogs(t, client, internal.DefaultNamespace, pod.Name, internal.CollectorContainerName, 500)
-			assert.NoError(c, err)
-			assert.Contains(c, podLogs, "Starting target allocator discovery", "Collector failed to start target allocator discovery. Received logs: %v", podLogs)
-
-			if i > 0 {
-				combinedPodLogs.WriteString("\n")
-			}
-			combinedPodLogs.WriteString(fmt.Sprintf("%v\n%v", pod.Name, podLogs))
-		}
-		assert.True(c, containsReadyAgentPod, "No OTel Collector agent pod found ready")
-		// NOTE: The target allocator distributes scrape jobs across agents when there are more than one.
-		// Compile all logs from agents first, then ensure that altogether they have the required logs.
-		assert.Regexp(c, serviceMonitorRegex, combinedPodLogs.String(), "Collector failed to start scrape job for serviceMonitor. Received logs: %v", combinedPodLogs.String())
-		assert.Regexp(c, podMonitorRegex, combinedPodLogs.String(), "Collector failed to start scrape job for podMonitor. Received logs: %v", combinedPodLogs.String())
-	}, 3*time.Minute, 3*time.Second, "Failed to find required agent pod logs")
-}
-
 // Internal telemetry metrics are only sent when an event occurs. Due to cluster
 // setup and potentially different events occurring on the cluster before the
 // test runs, different internal telemetry metrics will be sent. This method
@@ -1297,111 +1081,4 @@ func waitForAllNamespacesToBeCreated(t *testing.T, client *kubernetes.Clientset)
 		}
 		return true
 	}, 5*time.Minute, 10*time.Second)
-}
-
-// metricMatchFn decides whether a given metric (with its resource attributes)
-// should be counted. Return true to accept the metric.
-type metricMatchFn func(resAttrs pcommon.Map, metric pmetric.Metric) bool
-
-func checkMetricsAreEmitted(t *testing.T, mc *consumertest.MetricsSink, metricNames []string) {
-	checkMetrics(t, mc, metricNames, "", nil)
-}
-
-func checkMetricsFromApp(t *testing.T, mc *consumertest.MetricsSink, sdkLanguage, serviceName string, metricNames []string) {
-	checkMetrics(t, mc, metricNames, sdkLanguage+"/"+serviceName, func(resAttrs pcommon.Map, metric pmetric.Metric) bool {
-		if hasAttrMatch(resAttrs, "telemetry.sdk.language", sdkLanguage) && hasAttrMatch(resAttrs, "service.name", serviceName) {
-			return true
-		}
-		return metricDataPointsHaveAttrs(metric, "telemetry.sdk.language", sdkLanguage, "service.name", serviceName)
-	})
-}
-
-func checkMetrics(t *testing.T, mc *consumertest.MetricsSink, metricNames []string, label string, match metricMatchFn) {
-	metricsToFind := map[string]bool{}
-	for _, name := range metricNames {
-		metricsToFind[name] = false
-	}
-	require.Eventuallyf(t, func() bool {
-		for _, m := range mc.AllMetrics() {
-			for i := 0; i < m.ResourceMetrics().Len(); i++ {
-				rm := m.ResourceMetrics().At(i)
-				resAttrs := rm.Resource().Attributes()
-				for j := 0; j < rm.ScopeMetrics().Len(); j++ {
-					sm := rm.ScopeMetrics().At(j)
-					for k := 0; k < sm.Metrics().Len(); k++ {
-						metric := sm.Metrics().At(k)
-						if match == nil || match(resAttrs, metric) {
-							metricsToFind[metric.Name()] = true
-						}
-					}
-				}
-			}
-		}
-		var stillMissing, found []string
-		for _, name := range metricNames {
-			if metricsToFind[name] {
-				found = append(found, name)
-			} else {
-				stillMissing = append(stillMissing, name)
-			}
-		}
-		if label != "" {
-			t.Logf("[%s] found=%d missing=%d (%s)", label, len(found), len(stillMissing), strings.Join(stillMissing, ", "))
-		} else {
-			t.Logf("found=%d missing=%d (%s)", len(found), len(stillMissing), strings.Join(stillMissing, ", "))
-		}
-		return len(stillMissing) == 0
-	}, 3*time.Minute, 10*time.Second,
-		"failed to receive all metrics %s in 3 minutes", label)
-}
-
-func hasAttrMatch(attrs pcommon.Map, key, expected string) bool {
-	v, ok := attrs.Get(key)
-	return ok && v.Str() == expected
-}
-
-// metricDataPointsHaveAttrs checks whether any data point in a metric carries
-// all of the given key/value pairs. Pairs are passed as alternating key, value strings.
-func metricDataPointsHaveAttrs(metric pmetric.Metric, kvPairs ...string) bool {
-	check := func(attrs pcommon.Map) bool {
-		for i := 0; i < len(kvPairs)-1; i += 2 {
-			if !hasAttrMatch(attrs, kvPairs[i], kvPairs[i+1]) {
-				return false
-			}
-		}
-		return true
-	}
-	switch metric.Type() {
-	case pmetric.MetricTypeGauge:
-		for i := 0; i < metric.Gauge().DataPoints().Len(); i++ {
-			if check(metric.Gauge().DataPoints().At(i).Attributes()) {
-				return true
-			}
-		}
-	case pmetric.MetricTypeSum:
-		for i := 0; i < metric.Sum().DataPoints().Len(); i++ {
-			if check(metric.Sum().DataPoints().At(i).Attributes()) {
-				return true
-			}
-		}
-	case pmetric.MetricTypeHistogram:
-		for i := 0; i < metric.Histogram().DataPoints().Len(); i++ {
-			if check(metric.Histogram().DataPoints().At(i).Attributes()) {
-				return true
-			}
-		}
-	case pmetric.MetricTypeSummary:
-		for i := 0; i < metric.Summary().DataPoints().Len(); i++ {
-			if check(metric.Summary().DataPoints().At(i).Attributes()) {
-				return true
-			}
-		}
-	case pmetric.MetricTypeExponentialHistogram:
-		for i := 0; i < metric.ExponentialHistogram().DataPoints().Len(); i++ {
-			if check(metric.ExponentialHistogram().DataPoints().At(i).Attributes()) {
-				return true
-			}
-		}
-	}
-	return false
 }
